Eliminate hardcoded elements of bps report layout
The format/layout of bps report was hardcoded. There were hardcoded
integers for numbers of hyphens to print and lengths of fields, etc. I
rewrote the functions reponsible for displaying the report results with
the Table class from 'astropy.table'.  As a result, I could get rid of
the hardcoded elements of the layout and use formatting capabilities
provided by this class.