Simplify initializing number of retries
There are two default values for the number of retries in ctrl_bps: None
(no retries) and 5 (when memory auto scaling is enabled).  However, they
were not set consistently.  The first was set during creation of a
generic workflow job the other in the HTCondor plugin.  As the second
default value is also supposed to be used by all plugins, I made changes
to make ctrl_bps use in the same place, when creating a generic workflow
job.