Persist DAGMan job global id
When a user provides submission directory as a run id to bps report, it
retrieves the required information from various HTCondor files in the
submission directory.  However, none of this files contains the global
job id of the DAGMan job (at least in ver. 8.8 and 9.0 of HTCondor).
Hence I added an additional step to HTCondor submission which persists
the global id in a JSON file for future reference.