Modify the memory scaling mechanism
When memory scaling is enabled a job is removed from the job queue as
soon as the amount of memory it requests exceeds a given memory limit.
I altered the memory scaling mechanism to make it run a job one more
time in such a situation (if new run attempts are permitted) with the
memory allowable by the limit and remove the job from the queue
afterwards it fails due to insufficient memory anyway.